Effective-date-parser/requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867338309/work
aiohttp==3.8.0
aiosignal==1.2.0
altair==4.1.0
argcomplete==1.12.3
argon2-cffi==21.1.0
async-timeout==4.0.0a3
asynctest==0.13.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.0
chardet @ file:///C:/ci/chardet_1607706910910/work
charset-normalizer==2.0.7
click==8.0.3
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
configparser==5.0.2
coverage @ file:///C:/ci/coverage_1614614910274/work
cryptography @ file:///C:/ci/cryptography_1633502396740/work
Cython @ file:///C:/ci/cython_1626257059178/work
datasets==1.14.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.3.2
frozenlist==1.2.0
fsspec==2021.10.1
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.2
google-auth-oauthlib==0.4.6
grpcio==1.41.1
huggingface-hub==0.0.19
idna==3.3
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916842152/work
ipykernel==6.4.2
ipython==7.29.0
ipython_genutils==0.2.0
ipywidgets==7.6.5
jedi @ file:///C:/ci/jedi_1611315820849/work
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Markdown @ file:///C:/ci/markdown_1614364005059/work
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
multidict==5.2.0
multiprocess==0.70.12.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
notebook==6.4.5
numpy==1.21.3
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
olefile==0.46
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathtools==0.1.2
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow==8.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.21
protobuf==3.19.1
psutil==5.8.0
pyarrow==6.0.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.7.1
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
PyJWT @ file:///C:/ci/pyjwt_1619651810943/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywinpty==1.1.5
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
seqeval==1.2.2
shortuuid==1.0.1
simpletransformers==0.62.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
streamlit==1.1.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.0.4
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wandb==0.12.6
watchdog==2.1.6
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
xxhash==2.0.2
yarl==1.7.0
yaspin==2.1.0
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
